is_differentiable accounts for grad method
